pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fibrum-pdf"
version = "1.1.0"
description = "A fast PDF extractor; a 500 pages/s alternative to Marker, Docling, PyMUPDF4LLM & others."
authors = [{ name = "Adit Bajaj" }]
readme = "README.md"
requires-python = ">=3.9"
license = "AGPL-3.0"
dependencies = [
    "pydantic",
    "ijson"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "typer",
    "build",
]
benchmark = [
    "datasets",
    "docling",
    "altair",
    "apted",
    "lxml",
    "mistune",
    "polars",
    "pymupdf",
    "pymupdf4llm",
    "python-Levenshtein",
    "rapidfuzz",
    "rich",
    "scipy",
    "typer",
    "fibrum-pdf"
]

[project.urls]
Homepage = "https://github.com/intercepted16/fibrumpdf"
Repository = "https://github.com/intercepted16/fibrumpdf"
Issues = "https://github.com/intercepted16/fibrumpdf/issues"

[project.scripts]
fibrum-pdf = "fibrum_pdf.main:main"

[tool.cibuildwheel]
skip = ["*-musllinux*", "*i686*", "*aarch64*"]
build = [
    "cp3{9,10,11,12,13,14}-manylinux_x86_64",
    "cp3{9,10,11,12,13,14}-macosx_arm64"
]




[tool.cibuildwheel.linux]
environment = { LD_LIBRARY_PATH = "$PWD/lib/mupdf:$LD_LIBRARY_PATH", PATH = "/usr/local/go/bin:$PATH" }
before-all = [
    "yum install -y wget tar || (apt-get update && apt-get install -y wget)",
    "wget -q https://go.dev/dl/go1.24.0.linux-amd64.tar.gz",
    "tar -C /usr/local -xzf go1.24.0.linux-amd64.tar.gz",
    "/usr/local/go/bin/go version",
    "mkdir -p lib/mupdf",
    "curl -L -o lib/mupdf/libmupdf.so https://github.com/intercepted16/mupdf-prebuilts/releases/download/mupdf-1.27.0/libmupdf.so",
    "curl -L -o lib/mupdf/libmupdf.so.27.0 https://github.com/intercepted16/mupdf-prebuilts/releases/download/mupdf-1.27.0/libmupdf.so.27.0",
    "chmod +x lib/mupdf/*.so*"
]
repair-wheel-command = "auditwheel repair -w {dest_dir} --plat manylinux_2_28_x86_64 {wheel}"


[tool.cibuildwheel.macos]
environment = { DYLD_LIBRARY_PATH = "$PWD/lib/mupdf:$DYLD_LIBRARY_PATH" }
before-all = [
    "which go || brew install go@1.24 || true",
    "go version",
    "mkdir -p lib/mupdf",
    "curl -L -o lib/mupdf/libmupdf.dylib https://github.com/intercepted16/mupdf-prebuilts/releases/download/mupdf-1.27.0/libmupdf.dylib",
    "chmod +x lib/mupdf/*.dylib*"
]
repair-wheel-command = """
DYLD_LIBRARY_PATH=$PWD/lib/mupdf delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}
"""
